+++ NEWS.md
@@ -2,3 +2,4 @@
 
  * Add targets functions.
  * Add functions to help with [CNAG](https://www.cnag.eu/) projects.
+ * Add new function for transcriptomics.